fix(core): Rename files with type in exts, Add adapters, Player resolver, and utils

+++ src/commands/clear.command.ts
@@ -1,11 +1,10 @@
 import { DMChannel, Message } from 'discord.js'
 import { Command, CommandOptions } from '@src/command'
 import { Service } from 'typedi'
-import { Logger } from '@utils/logger'
+import { LoggerUtil } from '@utils/logger.util'
 
 @Service()
 export default class extends Command {
-
     get options(): CommandOptions {
         return {
             name: 'clear',
@@ -19,6 +18,9 @@ export default class extends Command {
         }
     }
 
+    // clear [limit:int]
+    // lottery close [limit:int]
+
     public async run(message: Message, args: string[]): Promise<void> {
         if (message.channel instanceof DMChannel) {
             return
@@ -31,10 +33,11 @@ export default class extends Command {
             return
         }
 
-        message.channel.bulkDelete(limit + 1).catch((error) => {
-            Logger.info('Error deleting messages', { error: error })
-            message.reply(`Erro a eliminar mensagens.`)
-        })
+        if (!message.channel.partial) {
+            message.channel.bulkDelete(limit + 1).catch((error) => {
+                LoggerUtil.info('Error deleting messages', { error: error })
+                message.reply(`Erro a eliminar mensagens.`)
+            })
+        }
     }
-
 }
